requirements.txt
discord.py==1.7.1
mysql-connector-python==8.0.24
Pillow==8.2.0
requests==2.25.1